# processor/stream: Re-introduce queue is full