## Motivation/summary

This patch re-introduces the `publish.QueueFull` error on the stream
processor layer. Allowing the APM Server to push back back when the
event output rate is significantly slower than the received event rate.

Since the semaphore is bound to 200 by default, this behavior should
only manifest when the concurrent number of in-flight requests exceeds
200 and Elasticsearch isn't able to process the APM Server's events as
fast as APM Server receives them.

The resulting behavior is slightly different than before since the back-
off will occur not on any bounded queue, but instead takes into account
the currently open requests that are being decoded.

This results in a very effective and relatively fast circuit breaker
when APM Server is overwhelmed and should prevent OOM issues when the
received event rate wildly exceeds the output event rate.

There are some open questions to this approach:

1. Is `1s` an acceptable "timeout" period?
  1.1. Should it be configurable?
2. Is `200` a good limit to set for the open and decoding requests?
  2.1. Would a lower number reduce the contention on `activeMu` lock and
       provide better performance?

## Checklist

<!--
Delete irrelevant items. The changelog should only be updated for user-facing changes.
Once the PR is ready for review there should be no unticked boxes.
-->

- [ ] Update [CHANGELOG.asciidoc](https://github.com/elastic/apm-server/blob/main/CHANGELOG.asciidoc)
~- [ ] Update [package changelog.yml](https://github.com/elastic/apm-server/blob/main/apmpackage/apm/changelog.yml) (only if changes to `apmpackage` have been made)~
- [ ] Documentation has been updated
- [ ] Benchmark change

## How to test these changes

TODO

## Related issues

Part of #7872