@marclop 2022-05-20
I ran the benchmarks again and the difference is really negligible, the variance I'd attribute to my own not extremely idle machine (tried to make it as idle as possible though):

```
name                                   old time/op    new time/op    delta
BackendProcessorParallelSemSize/1-100    6.82ms ± 1%    6.88ms ± 1%   ~     (p=0.400 n=3+3)
BackendProcessorParallelSemSize/2-100    4.18ms ± 0%    4.21ms ± 0%   ~     (p=0.100 n=3+3)
BackendProcessorParallelSemSize/4-100    3.11ms ± 4%    3.05ms ± 1%   ~     (p=0.700 n=3+3)
BackendProcessorParallelSemSize/8-100    2.85ms ± 0%    2.87ms ± 0%   ~     (p=0.100 n=3+3)

name                                   old speed      new speed      delta
BackendProcessorParallelSemSize/1-100  58.5MB/s ± 1%  58.0MB/s ± 1%   ~     (p=0.400 n=3+3)
BackendProcessorParallelSemSize/2-100  95.5MB/s ± 0%  94.7MB/s ± 0%   ~     (p=0.100 n=3+3)
BackendProcessorParallelSemSize/4-100   129MB/s ± 4%   131MB/s ± 1%   ~     (p=0.700 n=3+3)
BackendProcessorParallelSemSize/8-100   140MB/s ± 0%   139MB/s ± 0%   ~     (p=0.100 n=3+3)

name                                   old alloc/op   new alloc/op   delta
BackendProcessorParallelSemSize/1-100    13.8MB ± 0%    13.8MB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessorParallelSemSize/2-100    13.8MB ± 0%    13.8MB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessorParallelSemSize/4-100    13.8MB ± 0%    13.8MB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessorParallelSemSize/8-100    13.8MB ± 0%    13.7MB ± 0%   ~     (p=0.100 n=3+3)

name                                   old allocs/op  new allocs/op  delta
BackendProcessorParallelSemSize/1-100     24.8k ± 0%     24.8k ± 0%   ~     (p=0.300 n=3+3)
BackendProcessorParallelSemSize/2-100     24.8k ± 0%     24.8k ± 0%   ~     (p=1.000 n=3+3)
BackendProcessorParallelSemSize/4-100     24.8k ± 0%     24.8k ± 0%   ~     (p=1.000 n=3+3)
BackendProcessorParallelSemSize/8-100     24.9k ± 0%     24.9k ± 0%   ~     (p=1.000 n=3+3)
```