@marclop 2022-08-25
I'm closing this PR off for now since it's uncertain if we want to go this route.